broke data mapper to sqlite.
DataMapper expects Sqlite adapter, but DataObject expects Sqlite3.

There is code in dm-sqlite-adapter to change the adapter from sqlite to sqlite3, and the dm-do-adapter to create URI using adapter instead of scheme….

but that breaks JNDI case, when we want to scheme to stay 'java', but adapter to be whatever the underlying adapter is.

consolidated more of the logic that maps DM constants to DO constants

and changing DriverDefinition's verifyScheme to check for both scheme and jdbcScheme (in the case of 'sqlite' vs. 'sqlite3').